+++ src/components/search/SinopiaSearchResults.jsx
@@ -5,13 +5,11 @@ import React, { useState, useEffect } from 'react'
 import { connect } from 'react-redux'
 import { bindActionCreators } from 'redux'
 import PropTypes from 'prop-types'
-import Config from 'Config'
 import { getCurrentUser } from 'authSelectors'
 import { retrieveResource } from 'actionCreators/resources'
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import { faCopy, faEdit } from '@fortawesome/free-solid-svg-icons'
 import { findResource, findErrors } from 'selectors/resourceSelectors'
 import Alerts from '../Alerts'
+import SearchResultRows from './SearchResultRows'
 import SinopiaSort from './SinopiaSort'
 import _ from 'lodash'
 
@@ -20,11 +18,6 @@ export const searchRetrieveErrorKey = 'searchresource'
 
 const SinopiaSearchResults = (props) => {
   const [navigateEditor, setNavigateEditor] = useState(false)
-
-  const groupName = (uri) => {
-    const groupSlug = uri.split('/')[4]
-    return Config.groupsInSinopia[groupSlug]
-  }
 
   const handleCopy = (resourceURI) => {
     props.retrieveResource(props.currentUser, resourceURI, searchRetrieveErrorKey, true).then((success) => {
@@ -44,41 +37,6 @@ const SinopiaSearchResults = (props) => {
       props.history.push('/editor')
     }
   })
-
-  // Generates an HTML row
-  // TODO: Turn this function into a functional component
-  const generateRows = () => {
-    const rows = []
-    props.searchResults.forEach((row) => {
-      rows.push(<tr key={row.uri}>
-        <td>{ row.label }</td>
-        <td>
-          <ul className="list-unstyled">
-            { row.type?.map(type => <li key={type}>{type}</li>) }
-          </ul>
-        </td>
-        <td>{ groupName(row.uri) }</td>
-        <td><relative-time datetime={ row.modified }>{ row.modified }</relative-time></td>
-        <td>
-          <div className="btn-group" role="group" aria-label="Result Actions">
-            <button className="btn btn-link"
-                    title="Edit"
-                    onClick={e => handleEdit(row.uri, e) }>
-              <FontAwesomeIcon icon={faEdit} size="2x" />
-            </button>
-            <button type="button"
-                    className="btn btn-link"
-                    onClick={() => handleCopy(row.uri)}
-                    title="Copy"
-                    aria-label="Copy this resource">
-              <FontAwesomeIcon icon={faCopy} size="2x" />
-            </button>
-          </div>
-        </td>
-      </tr>)
-    })
-    return rows
-  }
 
   if (props.searchResults.length === 0) {
     return null
@@ -110,7 +68,7 @@ const SinopiaSearchResults = (props) => {
               </tr>
             </thead>
             <tbody>
-              { generateRows() }
+              <SearchResultRows searchResults={props.searchResults} handleEdit={handleEdit} handleCopy={handleCopy} />
             </tbody>
           </table>
         </div>
